Implemented skip/mark/reset for ByteArrayInputStream.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@71 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/io/AbstractInputStream.h
@@ -28,7 +28,7 @@ namespace sk {
         char read();
         std::vector<char> read(int number);
         std::vector<char>& read(std::vector<char>& buffer, int number);
-        long long skip(long long number);
+        int skip(int number);
         int available() const;
         bool markSupported() const;
         void mark(int readlimit);
